requirements.txt
# requirements.txt

Flask==2.3.3
Werkzeug==2.3.8
numpy==1.26.4
torch==2.1.2
torchvision==0.16.2
sentence-transformers==2.7.0
transformers==4.41.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==0.8.11